+++ tests/testthat/helper-rlang.R
@@ -43,3 +43,19 @@ skip_if_stale_backtrace <- local({
     skip_if(has_stale_backtrace)
   }
 })
+
+Rscript <- function(args, ...) {
+  out <- suppressWarnings(system2(
+    file.path(R.home("bin"), "Rscript"),
+    args,
+    ...,
+    stdout = TRUE,
+    stderr = TRUE,
+    timeout = 3
+  ))
+
+  list(
+    out = unstructure(out),
+    status = attr(out, "status")
+  )
+}